Rework of the Makefile for clarity and correctness
Fixed a bunch of dependencies, as well as added immediate status for binaries
that don't need to be rebuild every time that was forcing mapdata to be
regenerated unnecessarily.

Moved the most interesting variables to the top of the makefile and commented
everything.

Still need to fix the DEB and RPM targets though.